fix: fix client package
