+++ mirix/agent/agent.py
@@ -43,7 +43,7 @@ from mirix.llm_api.llm_api_tools import create
 from mirix.llm_api.llm_client import LLMClient
 from mirix.memory import summarize_messages
 from mirix.schemas.user import User
-from mirix.orm.enums import ToolType
+from mirix.schemas.enums import ToolType
 from mirix.schemas.agent import AgentState, AgentStepResponse, UpdateAgent
 from mirix.schemas.block import BlockUpdate
 from mirix.schemas.embedding_config import EmbeddingConfig
