+++ mirix/helpers/json_helpers.py
@@ -28,13 +28,13 @@ def parse_json(string) -> dict:
         result = json_loads(string)
         return result
     except Exception as e:
-        logger.debug("Error parsing json with json package: %s", e)
+        logger.debug(f"Error parsing json with json package: {e}")
 
     try:
         result = demjson.decode(string)
         return result
     except demjson.JSONDecodeError as e:
-        logger.debug("Error parsing json with demjson package: %s", e)
+        logger.debug(f"Error parsing json with demjson package: {e}")
 
     try:
         from json_repair import repair_json
@@ -43,5 +43,5 @@ def parse_json(string) -> dict:
         return result
 
     except Exception as e:
-        logger.debug("Error repairing json with json_repair package: %s", e)
+        logger.debug(f"Error repairing json with json_repair package: {e}")
         raise e
