+++ mirix/schemas/enums.py
@@ -1,6 +1,18 @@
 from enum import Enum
 
 
+class ToolType(str, Enum):
+    """Types of tools in Mirix"""
+    CUSTOM = "custom"
+    MIRIX_CORE = "mirix_core"
+    MIRIX_CODER_CORE = "mirix_coder_core"
+    MIRIX_MEMORY_CORE = "mirix_memory_core"
+    MIRIX_EXTRA = "mirix_extra"
+    MIRIX_MCP = "mirix_mcp"
+    MIRIX_MULTI_AGENT_CORE = "mirix_multi_agent_core"
+    USER_DEFINED = "user_defined"
+
+
 class ProviderType(str, Enum):
     anthropic = "anthropic"
