+++ mirix/schemas/sandbox_config.py
@@ -8,12 +8,9 @@ from pydantic import BaseModel, Field, model_validator
 from mirix.schemas.agent import AgentState
 from mirix.schemas.mirix_base import MirixBase, OrmMetadataBase
 from mirix.settings import tool_settings
-from mirix.log import get_logger
 
 
 # Sandbox Config
-
-logger = get_logger(__name__)
 class SandboxType(str, Enum):
     E2B = "e2b"
     LOCAL = "local"
@@ -73,6 +70,7 @@ class E2BSandboxConfig(BaseModel):
         Assign a default template value if the template field is not provided.
         """
         if data.get("template") is None:
+            # Use configured template if available
             data["template"] = tool_settings.e2b_sandbox_template_id
         return data
