+++ mirix/schemas/semantic_memory.py
@@ -6,7 +6,7 @@ from pydantic import Field, field_validator
 from mirix.constants import MAX_EMBEDDING_DIM
 from mirix.schemas.embedding_config import EmbeddingConfig
 from mirix.schemas.mirix_base import MirixBase
-from mirix.utils import get_utc_time
+from mirix.client.utils import get_utc_time
 
 
 class SemanticMemoryItemBase(MirixBase):
